+++ src/components/Popover/index.ts
@@ -1,3 +1,3 @@
-// Popover from copy-ui @ 2025-10-03.
+// Popover from copy-ui @ 2025-10-26.
 
-export { type Placement, Popover } from "./Popover";
+export { type Placement, Popover } from "./popover";