+++ src/components/Popover/CHANGELOG.md
@@ -9,3 +9,5 @@
 - 2025-09-25: Support `onOpen` and `onClose` props for `Popover.Trigger`'s
   render. And rename its `onClick` props to `onToggle`.
 - 2025-10-03: Fix missing dependencies in useEffect hooks.
+- 2025-10-26: Using `Modal` component in `Popover.Portal` to handle the portal
+  rendering.